--- backend/ModuMessenger/src/main/java/com/example/modumessenger/auth/CustomAuthenticationFailureHandler.java
@@ -1,27 +1,0 @@
-package com.example.modumessenger.auth;
-
-import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-@RequiredArgsConstructor
-public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        Logger log = LoggerFactory.getLogger("Authentication Failure");
-        log.error(exception.getMessage());
-
-        response.addHeader("token", null);
-        response.addHeader("userId", null);
-    }
-}
