--- backend/ModuMessenger/src/main/java/com/example/modumessenger/auth/JwtTokenProvider.java
@@ -1,108 +1,0 @@
-package com.example.modumessenger.auth;
-
-import io.jsonwebtoken.*;
-import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.annotation.PostConstruct;
-import java.util.*;
-
-@Component
-@RequiredArgsConstructor
-public class JwtTokenProvider {
-
-    private static Logger log = LoggerFactory.getLogger(JwtTokenProvider.class);
-
-    @Value("${security.jwt.token.secret-key}")
-    private String secretKey;
-    @Value("${security.jwt.token.expire-length:200000}")
-    private long validityInMilliseconds;
-
-    @Value("${security.jwt.token.access-expired}")
-    private Long ACCESS_TOKEN_EXPIRED_TIME;
-
-    @Value("${security.jwt.token.refresh-expired}")
-    private Long REFRESH_TOKEN_EXPIRED_TIME;
-
-    @PostConstruct
-    protected void init() {
-        secretKey = Base64.getEncoder().encodeToString(secretKey.getBytes());
-    }
-
-    public String createJwtAccessToken(String userId, List<String> roles) {
-        Claims claims = Jwts.claims().setSubject(userId);
-        claims.put("roles", roles);
-        Date now = new Date();
-        Date expiration = new Date(now.getTime() + ACCESS_TOKEN_EXPIRED_TIME);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(expiration)
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
-    }
-
-    public String createJwtRefreshToken(List<String> roles) {
-        Claims claims = Jwts.claims();
-        claims.put("uuid", UUID.randomUUID());
-        claims.put("roles", roles);
-        Date now = new Date();
-        Date expiration = new Date(now.getTime() + REFRESH_TOKEN_EXPIRED_TIME);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(expiration)
-                .signWith(SignatureAlgorithm.HS256, secretKey)
-                .compact();
-    }
-
-    public Claims getClaimsFromJwtToken(String jwtToken) {
-        try {
-            return Jwts.parser()
-                    .setSigningKey(secretKey)
-                    .parseClaimsJws(jwtToken)
-                    .getBody();
-
-        } catch (ExpiredJwtException e) {
-            return e.getClaims();
-        }
-    }
-
-    public String getRefreshToKenUUID(String refreshToken) {
-        return getClaimsFromJwtToken(refreshToken).get("uuid").toString();
-    }
-
-    public String findUserIdByJwt(String token) {
-        return getClaimsFromJwtToken(token).getSubject();
-    }
-
-    public List<String> getAuthentication(String token) {
-        return (List<String>) getClaimsFromJwtToken(token).get("roles");
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser().setSigningKey(secretKey).parseClaimsJws(token);
-            return true;
-        } catch (SignatureException ex) {
-            log.error("Invalid JWT signature");
-        } catch (MalformedJwtException ex) {
-            log.error("Invalid JWT token");
-        } catch (ExpiredJwtException ex) {
-            log.error("Expired JWT token");
-        } catch (UnsupportedJwtException ex) {
-            log.error("Unsupported JWT token");
-        } catch (IllegalArgumentException ex) {
-            log.error("JWT claims string is empty.");
-        } catch (NullPointerException ex) {
-            log.error("JWT RefreshToken is empty");
-        }
-        return false;
-    }
-}
-
