--- backend/ModuMessenger/src/main/java/com/example/modumessenger/auth/dto/TokenResponseDto.java
@@ -1,11 +1,0 @@
-package com.example.modumessenger.auth.dto;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class TokenResponseDto {
-    private String accessToken;
-    private String refreshToken;
-}
