--- backend/ModuMessenger/src/main/java/com/example/modumessenger/auth/repository/RefreshTokenRedisRepository.java
@@ -1,8 +1,0 @@
-package com.example.modumessenger.auth.repository;
-
-import com.example.modumessenger.auth.entity.RefreshToken;
-import org.springframework.data.repository.CrudRepository;
-
-public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, String> {
-}
-
