+++ backend/auth-service/src/main/java/com/example/authservice/auth/config/SecurityConfig.java
@@ -1,0 +1,77 @@
+package com.example.authservice.auth.config;
+
+import com.example.authservice.auth.jwt.JwtTokenProvider;
+import com.example.authservice.security.filter.CustomAuthenticationFilter;
+import com.example.authservice.security.handler.CustomAuthenticationFailureHandler;
+import com.example.authservice.security.handler.CustomAuthenticationSuccessHandler;
+import com.example.authservice.security.provider.CustomAuthenticationProvider;
+import com.example.authservice.service.RefreshTokenService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@Configuration
+@EnableWebSecurity
+@RequiredArgsConstructor
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    private final JwtTokenProvider jwtTokenProvider;
+    private final RefreshTokenService refreshTokenService;
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.authenticationProvider(customAuthenticationProvider());
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .httpBasic().disable()
+                .cors()
+                .and()
+                .csrf().disable()
+                .formLogin().disable()
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+
+        http.addFilterBefore(customAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+        ;
+    }
+
+    @Bean
+    @Override
+    public AuthenticationManager authenticationManager() throws Exception {
+        return super.authenticationManager();
+    }
+
+    public CustomAuthenticationSuccessHandler customAuthenticationSuccessHandler() throws Exception {
+        return new CustomAuthenticationSuccessHandler(jwtTokenProvider, refreshTokenService);
+    }
+
+    public CustomAuthenticationFailureHandler customAuthenticationFailureHandler() throws Exception {
+        return new CustomAuthenticationFailureHandler();
+    }
+
+    @Bean
+    public CustomAuthenticationFilter customAuthenticationFilter() throws Exception {
+        return new CustomAuthenticationFilter(authenticationManager(), customAuthenticationSuccessHandler(), customAuthenticationFailureHandler());
+    }
+
+    @Bean
+    public AuthenticationProvider customAuthenticationProvider() {
+        return new CustomAuthenticationProvider();
+    }
+
+    @Bean
+    public BCryptPasswordEncoder bCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+}
