+++ backend/auth-service/src/main/java/com/example/authservice/auth/controller/AuthController.java
@@ -1,0 +1,35 @@
+package com.example.authservice.auth.controller;
+
+import com.example.authservice.auth.dto.TokenResponseDto;
+import com.example.authservice.service.AccessTokenService;
+import com.example.authservice.service.RefreshTokenService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/auth")
+public class AuthController {
+
+    private final AccessTokenService accessTokenService;
+    private final RefreshTokenService refreshTokenService;
+
+    @PostMapping("/reissue")
+    public ResponseEntity<TokenResponseDto> reissue(@RequestHeader("Authorization") String accessToken, @RequestHeader("refresh-token") String refreshToken) {
+        return ResponseEntity.ok(refreshTokenService.reissueToken(accessToken, refreshToken));
+    }
+
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(@RequestHeader("Authorization") String accessToken) {
+        refreshTokenService.logoutToken(accessToken);
+        return ResponseEntity.ok().build();
+    }
+
+    @GetMapping("/check")
+    public ResponseEntity<String> checkAccessToken(@RequestHeader("Authorization") String accessToken) {
+        accessTokenService.checkToken(accessToken);
+        return ResponseEntity.ok("");
+    }
+}
+
