+++ backend/auth-service/src/main/java/com/example/authservice/auth/dto/TokenResponseDto.java
@@ -1,0 +1,15 @@
+package com.example.authservice.auth.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class TokenResponseDto {
+    private String accessToken;
+    private String refreshToken;
+}
+
