+++ backend/auth-service/src/main/java/com/example/authservice/auth/repository/RefreshTokenRedisRepository.java
@@ -1,0 +1,7 @@
+package com.example.authservice.auth.repository;
+
+import com.example.authservice.auth.entity.RefreshToken;
+import org.springframework.data.repository.CrudRepository;
+
+public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, String> {
+}
