+++ backend/auth-service/src/main/java/com/example/authservice/exception/ErrorResponse.java
@@ -1,0 +1,33 @@
+package com.example.authservice.exception;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+@Getter
+@Builder
+@RequiredArgsConstructor
+public class ErrorResponse {
+    private final HttpStatus status;
+    private final String code;
+    private final String message;
+
+    public ErrorResponse(ErrorCode errorCode) {
+        this.status = errorCode.getStatus();
+        this.code = errorCode.name();
+        this.message = errorCode.getMessage();
+    }
+
+    public static ResponseEntity<ErrorResponse> error(CustomException e) {
+        return ResponseEntity
+                .status(e.getErrorCode().getStatus())
+                .body(ErrorResponse.builder()
+                        .status(e.getErrorCode().getStatus())
+                        .code(e.getErrorCode().name())
+                        .message(e.getErrorTarget().isBlank() ?
+                                e.getErrorCode().getMessage() : e.getErrorCode().getMessage() + " ID: " + e.getErrorTarget())
+                        .build());
+    }
+}
