+++ backend/auth-service/src/main/java/com/example/authservice/member/dto/ProfileDto.java
@@ -1,0 +1,18 @@
+package com.example.authservice.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProfileDto implements Serializable {
+
+    private ProfileType profileType;
+    private String value;
+}
