+++ backend/auth-service/src/main/java/com/example/authservice/member/dto/ProfileType.java
@@ -1,0 +1,10 @@
+package com.example.authservice.member.dto;
+
+import lombok.Getter;
+
+@Getter
+public enum ProfileType {
+    PROFILE_STATUS_MESSAGE,
+    PROFILE_WALLPAPER,
+    PROFILE_IMAGE
+}
