+++ backend/auth-service/src/main/java/com/example/authservice/member/service/MemberService.java
@@ -1,0 +1,17 @@
+package com.example.authservice.member.service;
+
+import com.example.authservice.member.client.MemberFeignClient;
+import com.example.authservice.member.dto.MemberDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+
+    private final MemberFeignClient memberFeignClient;
+
+    public MemberDto getMember(String userId) {
+        return memberFeignClient.getMember(userId);
+    }
+}
