+++ backend/auth-service/src/main/java/com/example/authservice/security/dto/RequestLoginDto.java
@@ -1,0 +1,16 @@
+package com.example.authservice.security.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class RequestLoginDto {
+
+    private String userId;
+    private String email;
+}
