+++ backend/auth-service/src/main/resources/application.yml
@@ -1,18 +1,9 @@
 server:
-  port: 8080
+  port: 9900
 
 spring:
   application:
-    name: member-service
-
-  profiles:
-    default: local
-    include: oauth
-
-  config:
-    use-legacy-processing: true
-    activate:
-      on-profile: local
+    name: auth-service
 
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
@@ -20,20 +11,15 @@ spring:
     username: root
     password: root1234
 
+  redis:
+    host: localhost
+    port: 6379
+
   rabbitmq:
     host: localhost
     port: 5672
     username: guest
     password: guest
-
-  redis:
-    host: localhost
-    port: 6379
-
-  servlet:
-    multipart:
-      max-file-size: 100MB
-      max-request-size: 100MB
 
   jpa:
     hibernate:
@@ -45,27 +31,24 @@ spring:
         default_batch_fetch_size: 1000
     open-in-view: false
 
-  sleuth:
-    sampler:
-      probability: 1.0
-  zipkin:
-    base-url: http://localhost:9411
-
 logging:
   level:
     org.hibernate.SQL: debug
 
+security:
+  jwt:
+    token:
+      access-expired: 3600000
+      refresh-expired: 604800000
+      secret-key: 98FB6A50D8BE83F0521F6EDA48903D718DE1C41BF09D4D3671A5E3A6E4E0910A
+      expire-length: 3600000
+
 eureka:
   instance:
-    instance-id: member-service
+    instance-id: chat-service
 
   client:
     register-with-eureka: true
     fetch-registry: true
     service-url:
       defaultZone: http://localhost:8761/eureka
-
-project:
-  properties:
-    firebase-sdk-path: "firebase/modu_chat_firebase_service_key.json"
-    firebase-multicast-message-size: 500
