+++ backend/auth-service/src/test/java/com/example/authservice/AuthServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.authservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuthServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
